requirements-dev.txt
pytest
python-dotenv
flake8
pylint
pre-commit